I started making a list of comments, but decided just to make the changes directly, for the sake of your convenience.  I hope it is a convenience.  See the attached document.
I assigned new numbers as needed, and incremented version numbers where missed, after comparing against the last full draft (March 12).  Also, I fixed a few typos that I noticed.  Additionally, I fixed Theorem [113-002], which in this version for some reason incorrectly referenced continuants rather than occurrents.

The new text for process profiles added between pp. 69 and 85 needs reformatting to match the rest of the document, particularly with regard to sections numbers.  I did not attempt this.  However, I did restore the formatting and tracking numbering for the elucidation of process_profile on p. 82.  I also noticed that while temporal_part_of was initially defined on p. 63, it is also redefined in the added text on p. 71, presumably by virtue of a cut and paste operation.  These definitions are not identical.

Also with regard to temporal_part_of, I noticed that the new mereological axioms have not been applied to this relation, unlike the other two parthood relations.  Was this intentional, or an omission?

Mark